Move icons module
